Update handling of holiday values to "new" standard
This adapts parsing so that holidays can either be at the start or end
of a weekday range even in strict mode as per change of the OH
specification.

Additionally this handles holidays in a weekday range properly (fails in
strict mode, will add them to the end of the weekday range in non-strict
mode).

Note: to determine where the holidays go, that is at the end or at the beginning of the weekday range we store a flag, doing it this way is not exactly elegant.